fix up hog watcher service
