+++ frontend/src/scenes/hog-functions/misc/HogFunctionStatusIndicator.tsx
@@ -19,17 +19,7 @@ const displayMap: Record<HogWatcherState, DisplayOptions> = {
             </>
         ),
     },
-    [HogWatcherState.disabledForPeriod]: {
-        tagType: 'danger',
-        display: 'Disabled temporarily',
-        description: (
-            <>
-                The function has been disabled temporarily due to enough slow or failed requests. It will be re-enabled
-                soon.
-            </>
-        ),
-    },
-    [HogWatcherState.disabledIndefinitely]: {
+    [HogWatcherState.disabled]: {
         tagType: 'danger',
         display: 'Disabled',
         description: (
